Separated API classes into different files
