+++ lib/api/border.py
@@ -1,0 +1,9 @@
+import cherrypy
+from lib import queries
+
+
+@cherrypy.expose
+@cherrypy.tools.json_out()
+class Border(object):
+    def GET(self):
+        return queries.get_borders()
